requirements.txt

**requirements.txt**
```txt
torch
torchvision
torchaudio
transformers>=4.45.0
peft>=0.12.0
datasets
huggingface_hub
pillow
opencv-python-headless
accelerate
sentence-transformers
faiss-cpu
numpy
tqdm
kaggle-secrets  # for secret access (pre-installed on Kaggle)